+++ ee/connectors/msgcodec/messages.py
@@ -449,7 +449,7 @@ class PerformanceTrack(Message):
         self.used_js_heap_size = used_js_heap_size
 
 
-class StringDict(Message):
+class StringDictDeprecated(Message):
     __id__ = 50
 
     def __init__(self, key, value):
@@ -457,13 +457,30 @@ class StringDict(Message):
         self.value = value
 
 
-class SetNodeAttributeDict(Message):
+class SetNodeAttributeDictDeprecated(Message):
     __id__ = 51
 
     def __init__(self, id, name_key, value_key):
         self.id = id
         self.name_key = name_key
         self.value_key = value_key
+
+
+class StringDict(Message):
+    __id__ = 43
+
+    def __init__(self, key, value):
+        self.key = key
+        self.value = value
+
+
+class SetNodeAttributeDict(Message):
+    __id__ = 52
+
+    def __init__(self, id, name, value):
+        self.id = id
+        self.name = name
+        self.value = value
 
 
 class ResourceTimingDeprecated(Message):